addressing cr comments
