+++ src/converters/index.ts
@@ -1,3 +1,3 @@
-export * from './CommonConverters';
+export * from './RpcConverters';
 export * from './BindingConverters';
 export * from './RpcHttpConverters';
